requirements.txt
flask==2.3.3
flask-cors==4.0.0
ultralytics==8.0.196
deepface==0.0.79
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
psutil==5.9.0
